requirements.txt
flask>=3.0.0,<4.0.0
requests>=2.32.4,<3.0.0
beautifulsoup4>=4.13.4,<5.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.1.1,<2.0.0
lxml>=6.0.0,<7.0.0
gunicorn>=21.0.0,<22.0.0
